refactor url utils
