+++ reco_utils/dataset/movielens.py
@@ -7,7 +7,7 @@ import shutil
 import warnings
 import pandas as pd
 from zipfile import ZipFile
-from reco_utils.dataset.url_utils import maybe_download, download_path
+from reco_utils.dataset.download_utils import maybe_download, download_path
 from reco_utils.common.notebook_utils import is_databricks
 from reco_utils.common.constants import (
     DEFAULT_USER_COL,
